PathPlanAI-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.4
google-genai>=0.3.0
pdfplumber==0.11.0
python-docx==1.1.0
requests==2.31.0
groq==0.4.2